+++ valib/crc.h
@@ -33,16 +33,16 @@ class CRC
 {
 protected:
   uint32_t poly;
-  uint8_t  power;
+  unsigned power;
   uint32_t tbl[256];
   uint32_t tbl6[64];
 
 public:
   CRC() {};
-  CRC(uint32_t _poly, size_t _power)
+  CRC(uint32_t _poly, unsigned _power)
   { init(_poly, _power); };
 
-  void init(uint32_t poly, uint8_t power);
+  void init(uint32_t poly, unsigned power);
 
   /////////////////////////////////////////////////////////////////////////////
   // Pre- and post- CRC processing
